move: dto 갯수가 15개 넘어감에 따라 도메인별로 분류
dto 도메인별로 분류
