+++ src/main/java/gift/controller/OptionController.java
@@ -1,9 +1,9 @@
 package gift.controller;
 
-import gift.dto.OptionAddRequest;
-import gift.dto.OptionResponse;
-import gift.dto.OptionSubtractRequest;
-import gift.dto.OptionUpdateRequest;
+import gift.dto.option.OptionAddRequest;
+import gift.dto.option.OptionResponse;
+import gift.dto.option.OptionSubtractRequest;
+import gift.dto.option.OptionUpdateRequest;
 import gift.service.OptionService;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Pageable;
