--- src/main/java/gift/dto/KakaoAuthResponse.java
@@ -1,11 +1,0 @@
-package gift.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record KakaoAuthResponse(
-        @JsonProperty("kakao_account")
-        KakaoAccount kakaoAccount) {
-}
-
