+++ src/main/java/gift/dto/auth/AuthResponse.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.auth;
 
 public record AuthResponse(String token) {
     public static AuthResponse of(String token) {
