+++ src/main/java/gift/dto/auth/RegisterRequest.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
