+++ src/main/java/gift/dto/category/CategoryInformation.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.category;
 
 public record CategoryInformation(Long id, String name) {
     public static CategoryInformation of(Long id, String name) {
