+++ src/main/java/gift/dto/category/CategoryResponse.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.category;
 
 public record CategoryResponse(Long id, String name, String description, String color, String imageUrl) {
     public static CategoryResponse of(Long id, String name, String description, String color, String imageUrl) {
