+++ src/main/java/gift/dto/kakao/KakaoAuthInformation.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.kakao;
 
 public record KakaoAuthInformation(String name, String email) {
     public static KakaoAuthInformation of(String name, String email) {
