+++ src/main/java/gift/dto/kakao/KakaoAuthResponse.java
@@ -1,11 +1,11 @@
-package gift.dto;
+package gift.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record KakaoProfile(
-        @JsonProperty("nickname")
-        String name) {
+public record KakaoAuthResponse(
+        @JsonProperty("kakao_account")
+        KakaoAccount kakaoAccount) {
 }
