+++ src/main/java/gift/dto/kakao/KakaoAuthToken.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
