+++ src/main/java/gift/dto/kakao/KakaoProfile.java
@@ -1,13 +1,11 @@
-package gift.dto;
+package gift.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record KakaoAccount(
-        @JsonProperty("profile")
-        KakaoProfile profile,
-        @JsonProperty("email")
-        String email) {
+public record KakaoProfile(
+        @JsonProperty("nickname")
+        String name) {
 }
