+++ src/main/java/gift/dto/option/OptionResponse.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.option;
 
 public record OptionResponse(Long id, String name, Integer quantity) {
     public static OptionResponse of(Long id, String name, Integer quantity) {
