+++ src/main/java/gift/dto/product/ProductBasicInformation.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.product;
 
 public record ProductBasicInformation(Long id, String name, Integer price) {
     public static ProductBasicInformation of(Long id, String name, Integer price) {
