+++ src/main/java/gift/dto/wishproduct/WishProductResponse.java
@@ -1,4 +1,6 @@
-package gift.dto;
+package gift.dto.wishproduct;
+
+import gift.dto.product.ProductBasicInformation;
 
 public record WishProductResponse(Long id, ProductBasicInformation productBasicInformation, Integer quantity) {
     public static WishProductResponse of(Long id, ProductBasicInformation productBasicInformation, Integer quantity) {
