+++ src/main/java/gift/dto/wishproduct/WishProductUpdateRequest.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wishproduct;
 
 import jakarta.validation.constraints.PositiveOrZero;
