+++ src/main/java/gift/service/api/KakaoApiService.java
@@ -3,9 +3,9 @@ package gift.service.api;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.config.properties.KakaoProperties;
-import gift.dto.KakaoAuthInformation;
-import gift.dto.KakaoAuthResponse;
-import gift.dto.KakaoAuthToken;
+import gift.dto.kakao.KakaoAuthInformation;
+import gift.dto.kakao.KakaoAuthResponse;
+import gift.dto.kakao.KakaoAuthToken;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
