+++ src/main/java/gift/service/auth/AuthService.java
@@ -1,11 +1,11 @@
 package gift.service.auth;
 
 import gift.config.properties.JwtProperties;
-import gift.dto.AuthResponse;
-import gift.dto.KakaoAuthInformation;
-import gift.dto.KakaoAuthToken;
-import gift.dto.LoginRequest;
-import gift.dto.RegisterRequest;
+import gift.dto.auth.AuthResponse;
+import gift.dto.kakao.KakaoAuthInformation;
+import gift.dto.kakao.KakaoAuthToken;
+import gift.dto.auth.LoginRequest;
+import gift.dto.auth.RegisterRequest;
 import gift.exception.DuplicatedEmailException;
 import gift.exception.InvalidLoginInfoException;
 import gift.exception.NotFoundElementException;
