+++ src/test/java/gift/service/OptionServiceTest.java
@@ -1,8 +1,8 @@
 package gift.service;
 
-import gift.dto.OptionAddRequest;
-import gift.dto.OptionSubtractRequest;
-import gift.dto.OptionUpdateRequest;
+import gift.dto.option.OptionAddRequest;
+import gift.dto.option.OptionSubtractRequest;
+import gift.dto.option.OptionUpdateRequest;
 import gift.exception.BadRequestException;
 import gift.exception.DuplicatedNameException;
 import org.assertj.core.api.Assertions;
